Extract WorkspaceSwitcher & ContentArea components, fix flaky tooltip test
Move WorkspaceBubble, WorkspaceEditor, WORKSPACE_COLORS, and new
WorkspaceSwitcher wrapper out of sidebar.renderer.tsx into
features/workspaces/WorkspaceSwitcher.tsx. Move ContentArea from
Shell.tsx into components/ContentArea.tsx. Update design system page
imports and add editor/switcher previews. Fix flaky TooltipLayer test
by scoping vi.useFakeTimers to setTimeout/clearTimeout only — prevents
React 18 scheduling interference.
