+++ design-system/src/pages/components/web-content-host.mdx
@@ -3,7 +3,7 @@ title: Web Content Host
 ---
 
 import { LivePreview } from "../../components/LivePreview"
-import { ContentArea } from "@/Shell"
+import { ContentArea } from "@/components/ContentArea"
 
 # Web Content Host
