+++ design-system/tsconfig.json
@@ -23,9 +23,11 @@
     "../src/renderer/src/Shell.tsx",
     "../src/renderer/src/components/TooltipLayer.tsx",
     "../src/renderer/src/components/Icon.tsx",
+    "../src/renderer/src/components/ContentArea.tsx",
     "../src/features/**/*.renderer.tsx",
     "../src/features/**/*.store.ts",
     "../src/features/**/*.shared.ts",
-    "../src/features/command-palette/resolve-input.ts"
+    "../src/features/command-palette/resolve-input.ts",
+    "../src/features/workspaces/WorkspaceSwitcher.tsx"
   ]
 }
